test: split search orchestration recovery tests by concern
